// frontend/src/lib/generated/types.gen.ts
export enum DirectionEnum {
  LTR = "ltr",
  RTL = "rtl",
}

/**
 * An enum field
 */
export enum IconPositionEnum {
  BEFORE_TEXT = "before_text",
  AFTER_TEXT = "after_text",
}

/**
 * An enum field
 */
export enum AlignmentEnum {
  CENTER = "center",
  LANGUAGE_DIRECTION = "language-direction",
}